return type control added

+++ api/services/receipt.js
@@ -122,6 +122,8 @@ class ReceiptService {
 
 			let sum_y_vertices = [];
 			let sum_confidence;
+			let kdv_y_vertices = [];
+			let kdv_confidence;
 			let all_y_vertices = [];
 
 			fullTextAnnotation.pages.forEach(page => {
@@ -136,7 +138,14 @@ class ReceiptService {
 									all_y_vertices.push(-5000);
 								});
 							}
-							else if (wordText === '*') {
+							else if (wordText === 'KDV' || wordText === 'TOPKDV') {
+								kdv_confidence = word.confidence;
+								word.boundingBox.vertices.forEach(v => {
+									kdv_y_vertices.push(v.y);
+									all_y_vertices.push(-5000);
+								});
+							}
+							else if (wordText === '*' || wordText === '.' || wordText === '+') {
 								word.boundingBox.vertices.forEach(() => {
 									all_y_vertices.push(-5000);
 								});
@@ -151,6 +160,8 @@ class ReceiptService {
 				});
 			});
 
+			sum_y_vertices = sum_y_vertices.slice(-4);
+
 			const grouped_vertices = all_y_vertices.reduce((acc, curr, i) => {
 				if (i % 4 === 0) {
 					acc.push([ curr ]);
@@ -161,21 +172,29 @@ class ReceiptService {
 				return acc;
 			}, []);
         
-			let arr = [];
+			let arr_for_sum = [];
+			let arr_for_kdv = [];
 
 			for (let i = 0;i < grouped_vertices.length;i++) {
-				arr.push(getDifference(grouped_vertices[i], sum_y_vertices));
+				arr_for_sum.push(getDifference(grouped_vertices[i], sum_y_vertices));
+				arr_for_kdv.push(getDifference(grouped_vertices[i], kdv_y_vertices));
 			}
 
-			let closest = arr.reduce(function(prev, curr) {
+			let closest = arr_for_sum.reduce(function(prev, curr) {
+				return (Math.abs(curr - 0) < Math.abs(prev - 0) ? curr : prev);
+			});
+			
+			let closest_for_kdv = arr_for_kdv.reduce(function(prev, curr) {
 				return (Math.abs(curr - 0) < Math.abs(prev - 0) ? curr : prev);
 			});
 
-			let index_of_value = arr.indexOf(closest)+1;
+			let index_of_value = arr_for_sum.indexOf(closest)+1;
+			let index_of_kdv = arr_for_kdv.indexOf(closest_for_kdv)+1;
 
 			let l = 0;
 			let receipt_sum;
-
+			let receipt_kdv;
+	
 			fullTextAnnotation.pages.forEach(page => {
 				page.blocks.forEach(block => {
 					block.paragraphs.forEach(paragraph => {
@@ -184,6 +203,9 @@ class ReceiptService {
 							l = l + 1;
 							if (l === index_of_value) {
 								receipt_sum = wordText;
+							}
+							else if (l === index_of_kdv) {
+								receipt_kdv = wordText;
 							}
 						});
 					});
@@ -195,9 +217,13 @@ class ReceiptService {
 				return (((a[0]-b[0]) + (a[1]-b[1]) + (a[2]-b[2]) + (a[3]-b[3])) / 4);
 			}
 
+			if (isNaN(receipt_sum.replace(',', '.'))) {
+				return {};
+			}
+
 			fs.unlinkSync(targetPath);
 
-			return receipt_sum;
+			return {receipt_sum, receipt_kdv};
 		}
 		catch (err) {
 			console.log(err);
